Reeling Adjustments
Outside of goal circle is now the perfect zone for reeling, with clicks too early AND too late from the outer ring of the goal taking off from the score. New parameters are more intuitive and easier to adjust in inspector.

Sprite adjustments: goal circle is now just a circle and ring sprite is now a ring sprite instead of masked circles (more like how we will actually do it)